feat: 문자열을 나눠서 int 배열로 만드는 기능
